requirements.txt
telethon
python-dotenv
requests
aiohttp
loguru
fastapi
uvicorn
pydantic
psycopg2-binary
sqlalchemy
dbt-postgres

